changed all uses of BankTrustResponse to Bank (#89)
* changed all uses of BankTrustResponse to Bank

* changed port: String to  port: Int

* moved Bank class under common

* fee: Int to Double

* added trust parameter in fun bank()

* added trust parameter in fun bank()

* added correct account number to mock bank()

* Banks.kt --> BankList.kt

* Banks.kt --> BankList.kt

* Banks.kt --> BankList.kt

* Double to Int

* Double to Int

* ip_address to ipAddress

* edited emptyValidator

+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/validatordto/Validators.kt
@@ -32,7 +32,7 @@ data class Validator(
     val version: String,
 
     @SerialName("default_transaction_fee")
-    val defaultTransactionFee: Double,
+    val defaultTransactionFee: Int,
 
     @SerialName("root_account_file")
     val rootAccountFile: String,
@@ -44,7 +44,7 @@ data class Validator(
     val seedBlockIdentifier: String,
 
     @SerialName("daily_confirmation_rate")
-    val dailyConfirmationRate: Double? = null,
+    val dailyConfirmationRate: Int? = null,
 
     @SerialName("trust")
     val trust: Double
